Move CC check into actions
