+++ src/applications/vaos/new-appointment/newAppointmentFlow.js
@@ -1,30 +1,18 @@
-import * as Sentry from '@sentry/browser';
-import { captureError } from '../utils/error';
-
 import {
   getFormData,
   getNewAppointment,
   getEligibilityStatus,
   vaosCommunityCare,
-  getCCEType,
   getTypeOfCare,
-  selectSystemIds,
-  vaosVSPAppointmentNew,
 } from '../utils/selectors';
 import { FACILITY_TYPES, FLOW_TYPES, TYPES_OF_CARE } from '../utils/constants';
-import { getCommunityCare, getSitesSupportingVAR } from '../services/var';
 import {
   showTypeOfCareUnavailableModal,
   startDirectScheduleFlow,
   startRequestAppointmentFlow,
   updateFacilityType,
-  updateCCEnabledSystems,
-  updateCCEligibility,
+  checkCommunityCareEligibility,
 } from './redux/actions';
-import {
-  getOrganizations,
-  getSiteIdFromOrganization,
-} from '../services/organization';
 
 const AUDIOLOGY = '203';
 const SLEEP_CARE = 'SLEEP';
@@ -65,42 +53,6 @@ function isPodiatry(state) {
   return getFormData(state).typeOfCareId === PODIATRY;
 }
 
-async function checkCommunityCareEligibility(state, dispatch) {
-  const communityCareEnabled = vaosCommunityCare(state);
-  const useVSP = vaosVSPAppointmentNew(state);
-
-  if (!communityCareEnabled) {
-    return false;
-  }
-
-  try {
-    // Check if user registered systems support community care...
-    const siteIds = selectSystemIds(state);
-    const parentIds = await getOrganizations({ siteIds, useVSP });
-    const ccSites = await getSitesSupportingVAR(parentIds);
-    const ccEnabledSystems = parentIds.filter(parent =>
-      ccSites.some(site => site.id === getSiteIdFromOrganization(parent)),
-    );
-    dispatch(updateCCEnabledSystems(ccEnabledSystems, parentIds));
-
-    // Reroute to VA facility page if none of the user's registered systems support community care.
-    if (ccEnabledSystems.length) {
-      const response = await getCommunityCare(getCCEType(state));
-
-      dispatch(updateCCEligibility(response.eligible));
-
-      return response.eligible;
-    }
-  } catch (e) {
-    captureError(e);
-    Sentry.captureMessage(
-      'Community Care eligibility check failed with errors',
-    );
-  }
-
-  return false;
-}
-
 export default {
   home: {
     url: '/',
@@ -121,7 +73,7 @@ export default {
       } else if (isEyeCare(state)) {
         return 'typeOfEyeCare';
       } else if (isCommunityCare(state)) {
-        const isEligible = await checkCommunityCareEligibility(state, dispatch);
+        const isEligible = await dispatch(checkCommunityCareEligibility());
 
         if (isEligible && isPodiatry(state)) {
           // If CC enabled systems and toc is podiatry, skip typeOfFacility
@@ -177,7 +129,7 @@ export default {
 
       // check that the result does have a ccId
       if (getTypeOfCare(data)?.ccId) {
-        const isEligible = await checkCommunityCareEligibility(state, dispatch);
+        const isEligible = await dispatch(checkCommunityCareEligibility());
 
         if (isEligible) {
           return 'typeOfFacility';
