+++ src/applications/vaos/new-appointment/redux/actions.js
@@ -1,4 +1,5 @@
 import moment from 'moment';
+import * as Sentry from '@sentry/browser';
 
 import recordEvent from 'platform/monitoring/record-event';
 
@@ -19,6 +20,7 @@ import {
   getRootIdForChosenFacility,
   getSiteIdForChosenFacility,
   vaosVSPAppointmentNew,
+  getCCEType,
 } from '../../utils/selectors';
 import {
   getPreferences,
@@ -26,10 +28,13 @@ import {
   submitRequest,
   submitAppointment,
   sendRequestMessage,
+  getSitesSupportingVAR,
+  getCommunityCare,
 } from '../../services/var';
 import {
   getOrganizations,
   getIdOfRootOrganization,
+  getSiteIdFromOrganization,
 } from '../../services/organization';
 import { getLocation } from '../../services/location';
 import { getSupportedHealthcareServicesAndLocations } from '../../services/healthcare-service';
@@ -160,14 +165,6 @@ export function updateFormData(page, uiSchema, data) {
     page,
     uiSchema,
     data,
-  };
-}
-
-export function updateCCEnabledSystems(ccEnabledSystems, parentFacilities) {
-  return {
-    type: FORM_VA_SYSTEM_UPDATE_CC_ENABLED_SYSTEMS,
-    ccEnabledSystems,
-    parentFacilities,
   };
 }
 
@@ -620,10 +617,51 @@ export function openCommunityCarePreferencesPage(page, uiSchema, schema) {
   };
 }
 
-export function updateCCEligibility(isEligible) {
-  return {
-    type: FORM_UPDATE_CC_ELIGIBILITY,
-    isEligible,
+export function checkCommunityCareEligibility() {
+  return async (dispatch, getState) => {
+    const state = getState();
+    const communityCareEnabled = vaosCommunityCare(state);
+    const useVSP = vaosVSPAppointmentNew(state);
+
+    if (!communityCareEnabled) {
+      return false;
+    }
+
+    try {
+      // Check if user registered systems support community care...
+      const siteIds = selectSystemIds(state);
+      const parentFacilities = await getOrganizations({ siteIds, useVSP });
+      const ccSites = await getSitesSupportingVAR(
+        parentFacilities.map(parent => getSiteIdFromOrganization(parent)),
+      );
+      const ccEnabledSystems = parentFacilities.filter(parent =>
+        ccSites.some(site => site.id === getSiteIdFromOrganization(parent)),
+      );
+      dispatch({
+        type: FORM_VA_SYSTEM_UPDATE_CC_ENABLED_SYSTEMS,
+        ccEnabledSystems,
+        parentFacilities,
+      });
+
+      // Reroute to VA facility page if none of the user's registered systems support community care.
+      if (ccEnabledSystems.length) {
+        const response = await getCommunityCare(getCCEType(state));
+
+        dispatch({
+          type: FORM_UPDATE_CC_ELIGIBILITY,
+          isEligible: response.eligible,
+        });
+
+        return response.eligible;
+      }
+    } catch (e) {
+      captureError(e);
+      Sentry.captureMessage(
+        'Community Care eligibility check failed with errors',
+      );
+    }
+
+    return false;
   };
 }
